Allow to turn off the sicky mode by setting v-sticky="null" (#20)
* Allow to turn off the sicky mode by setting v-sticky="null"

Allow to turn off the sicky mode by setting v-sticky="null" and to
restore the sticky behaviour by setting a new configuration object.
Example page updated to show the feature

* Allow to turn off the sticky mode

Allow to turn off the sticky mode by setting a 'disabled' property to
the configuration object. Setting it to 'false' disable the sticky
behaviour. If set to 'true' or not set at all it enables it.

A class is added to the elment when in 'sticky' position and it is
removed when it is in its 'normal' position. The classe needs to
bespecified adding a propety 'className' to the configuration object.

Code has been improved to handle more than one sticky element present
in the same _page_.

Example app has been updated to showcase the code changes.

* Update README.md file to include 'disabled'

Update README.md file to include 'disabled' property to turn off
and on the sticky behaviour.